First pass at breaking out files
