+++ lambda/requirements.txt
@@ -1,0 +1,14 @@
+asn1crypto==0.24.0
+configparser==3.5.0b2
+cryptography==2.6.1
+entrypoints==0.3
+enum34==1.1.6
+ipaddress==1.0.17
+keyring==17.1.1
+keyrings.alt==3.1.1
+powerline-status==2.7
+pycrypto==2.6.1
+PyGObject==3.30.4
+pyxdg==0.25
+SecretStorage==2.3.1
+six==1.12.0
